Clean up DecodeNextFrame path to something sane
As seen in issue #21, the MP3 decoder would barf on a single lost sync
error, even though it was not actually a fatel error.

Rewrote the logic to include another Input() on DNF failure, and only die
when Input() actually fails.  Seems to fix the problem in issue #21 and also
did not break the WebRadio.