Preallocate all memory and move imdct_s to progmem
Because the mp3 player is a long living object and needs to allocate
large contiguous chunks on each frame, it is safer to allocate the
large xr[] and tmp[] arrays at the beginning.  Add them to the frame
structure which lives the entire time the mp3 object is alive.  No
heap allocations are now used by the mp3 player.

Move the imdct_s[] array to progmem, but do a memcpy_P() to stack
as the macros may cause a 16-bit access after optimization.  It's only
36 words long and each data is read, anyway, so this doesn't increase
the CPU work significantly.