Ensure BPS=16 is set in MP3 decoder
Valgrind found another uninitialized use.  BitsPerSample was not being
called by the LIBMAD decoder, so it was possible to start out in a bad
state.  Now always set it to 16, the decoder native output size.

Channels are initialized to 2, but this setting is always reset on
a per-frame basis anyway.