Remove esp32 build for now
The ESP32 CI system has moved to the GH CI, from TravisCI.  The repo
also uses Python2 and not Python3.  These two combine to make the
existing ESP32 CI scripts almost unrunnable on the TravisCI instance
we're using.

Drop them for now, but hopefully we'll be able to reintegrate them at
some point in the future