Make as much code common between AudioFileSourceHTTPStream and AudioFileSourceICYStream as possible
Most of the AudioFileSourceHTTPStream was repeated in
AudioFileSourceICYStream which caused some inconsistencies, e.g.
redirects were enabled in ICY unconditionally and in HTTP only on
devices other than ESP32.

Now just the ICY-specific parts are implemented in
AudioFileSourceICYStream. readInternal() is implemented only in
AudioFileSourceHTTPStream and it delegates actual stream reading and
parsing to a new virtual method parseInternal(). This wasy
AudioFileSourceHTTPStream can merely read the data into the buffer and
AudioFileSourceICYStream can parse metadata at will.