Scan MP3 for header, avoid multiple burst errors
MP3 comes in chunks of encoded data started by a header sequence.
Scan for it in the decoder and only pass in data that starts with
this header to avoid the burst errors

Avoids a avalance of `lost synchronization` error reports when there
is a file error.

Also enable some of the web sources on the RP2040/RP2350